Add Containerizing Apps PDF
This PDF includes the exercise steps for the related Global Summit session on Containerizing Apps with InterSystems IRIS.

The associated source code is not included here, but the steps of the exercise may be helpful for further exploring. To check out the full experience, register for the Containerizing Apps with InterSystems IRIS experience lab!